LPS-148078 Updated code link and scope
