+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/styles/pages/_team-members.scss
@@ -16,6 +16,9 @@
 	}
 }
 .cp-manager-product-button {
+	&:hover {
+		border-color: var(--color-action-neutral-default);
+	}
 	& span {
 		width: 12px;
 		margin-left: 6px;
